redirects, seperation of concern and imports
